Split out a class that takes care of running ZooKeeper server
Need this to simplify the way this will be done when we start supporting
dynamic reconfiguration of ZooKeeper cluster
